Convert blog to use Eleventy instead of Next.js
This came from me staring at the Next.js 13/14 update and just not
wanting to do it 🙈 Our blog is hella basic, so Next.js was somewhat
overkill either way. I've been playing around with Eleventy recently
and wanted to try converting the blog to it.

Overall the dependency management should be much simpler now. There is
a downside in that it introduces a new framework, but I'd say 11ty is
pretty basic to work with and given that we don't actively develop
on the blog I'd say its an okay trade-off.